Rename Android module
