+++ android/build.gradle
@@ -78,7 +78,7 @@ def configureReactNativePom(def pom) {
         name packageJson.title
         artifactId packageJson.name
         version = packageJson.version
-        group = "nabidreams.audiorecorder"
+        group = "nabidreams.media"
         description packageJson.description
         url packageJson.repository.baseUrl
