+++ android/src/main/java/nabidreams/media/AudioModule.kt
@@ -1,14 +1,14 @@
-package nabidreams.audiorecorder
+package nabidreams.media
 
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
 import com.facebook.react.bridge.Callback
 
-class AudioRecorderModule(private val reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext) {
+class AudioModule(private val reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext) {
 
     override fun getName(): String {
-        return "AudioRecorder"
+        return "Audio"
     }
 
     @ReactMethod
