+++ example/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.audiorecorderexample",
+    package = "nabidreams.media.audioexample",
 )
 
 android_resource(
     name = "res",
-    package = "com.audiorecorderexample",
+    package = "nabidreams.media.audioexample",
     res = "src/main/res",
 )
