+++ example/android/app/src/main/java/nabidreams/media/audioexample/MainApplication.java
@@ -1,4 +1,4 @@
-package nabidreams.audiorecorder.example;
+package nabidreams.media.audioexample;
 
 import android.app.Application;
 import android.content.Context;
