+++ example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">AudioRecorderExample</string>
+    <string name="app_name">AudioExample</string>
 </resources>
