+++ example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AudioRecorderExample",
-  "displayName": "AudioRecorderExample"
+  "name": "AudioExample",
+  "displayName": "AudioExample"
 }
